requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.1+cu118
torchvision==0.17.1+cu118
torchaudio==2.2.1+cu118
accelerate==0.27.2
clip==0.2.0
diffusers==0.30.2
lpips==0.1.4
matplotlib==3.9.2
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
packaging==24.1
Pillow==9.5.0
scipy==1.14.1
torchmetrics==1.3.2
traitlets==5.14.2
ipython==8.12.3
matplotlib-inline==0.1.7
transformers==4.38.2
